Use Long instead Int for entity ids

+++ app/src/main/java/com/andremion/hostel/app/home/CityViewModel.kt
@@ -28,7 +28,7 @@ class CityViewModel(
 
     val data = ObservableField<City>()
 
-    fun loadCity(city: Int) {
+    fun loadCity(city: Long) {
         addDisposable(hostelRepository.getById(city)
                 .observeOn(appSchedulers.main)
                 .subscribe(data::set))
